requirements.txt
pyyaml>=6.0
requests>=2.32
tzlocal>=5.3